requirements.txt
numpy 
pandas 
opencv-python
matplotlib
seaborn
tensorflow<2.11
keras<2.11
pyarrow
fastapi
uvicorn
sqlalchemy
python-jose
passlib